refactor: extract bp_memory+security+config+logs to routes/infra.py
Phase 5.11 of the incremental modularisation. Bundles four small Blueprints
(15 routes total) into a single file because each is only 3-4 routes and
they are logically adjacent observability concerns:

  bp_logs     (4) — /api/logs, /api/flow[-events], /api/logs-stream
  bp_memory   (4) — /api/memory[-files], /api/file, /api/memory-analytics
  bp_security (3) — /api/security/{threats,signatures,posture}
  bp_config   (4) — /api/llmfit, /api/cost-optimizer, /api/cost-optimization,
                     /api/automation-analysis

dashboard.py shrinks by ~700 lines. Module-level helpers stay in dashboard.py
and are reached from handlers via late ``import dashboard as _d``, matching
the established pattern. ``api_security_threats`` calls ``api_brain_history``
from routes.brain via late import.

Smoke-tested: all 15 extracted routes + /api/flow HEAD + SSE status probes +
12 cross-phase regression endpoints return 200. No errors in server log.
